+++ src/effect/ChainSharp.Effect/Models/WorkQueue/DTOs/CreateWorkQueue.cs
@@ -23,7 +23,7 @@ public class CreateWorkQueue
     /// <summary>
     /// Optional manifest ID when this entry was queued from a scheduled manifest.
     /// </summary>
-    public int? ManifestId { get; set; }
+    public long? ManifestId { get; set; }
 
     /// <summary>
     /// Dispatch priority. Range: 0-31 (clamped in WorkQueue.Create). Defaults to 0.
